[Refactor #12] Redis, RefreshToken common 모듈로 이동
