+++ customer/src/test/java/com/project/customer/user/service/UserServiceTest.java
@@ -2,6 +2,7 @@ package com.project.customer.user.service;
 
 import com.project.core.common.oauth.GoogleOauth;
 import com.project.core.common.oauth.dto.GoogleResponse;
+import com.project.core.common.token.RefreshTokenService;
 import com.project.core.common.token.TokenProvider;
 import com.project.core.domain.user.User;
 import com.project.core.domain.user.customer.CustomerDog;
